use editorial pkg product type
